// src/types/popularMoviesTypes.ts
export type PopularMovieType = {
   adult: boolean;
   backdrop_path: string;
   genre_ids: number[];
   id: number;
   original_language: string;
   original_title: string;
   overview: string;
   popularity: number;
   poster_path: string;
   release_date: string;
   title: string;
   video: boolean;
   vote_average: number;
   vote_count: number;
}

export type RootObject = {
   page: number;
   results: PopularMovieType[];
   total_pages: number;
   total_results: number;
}

export type homeStateType = {
   popularMovies: Array<PopularMovieType>;
   bannerMovie: PopularMovieType
}